My GCI experience
My participation in Google Code In

Google Code-in - a competition organized annually by Google, addressed to junior high school students, grades 7-8 of elementary school and high schools (aged 13-17), involving the performance of various tasks for open source organizations. The competition was originally called the Google Highly Open Participation Contest, but in 2010 it was slightly changed and organized under the name Google Code-in.Everyone will find something for themselves in this competition

I was supposed to take part also in that year, but I do not know why, even though I did not want to register. This year I found out about the competition on social media - I thought I would try my hand, especially since it was my last chance. I filled in the proper application and took part. I searched for coding tasks once (from about 4-5 years I'm interested in c ++ and general IT) , but currently I do not have a computer (I am at the stage of submitting a new part) and I had to let it go.
My first task was to create a simulation game on the Android in Pocket Code, I admit it did not go well for me. Then the first time I had contact with mentors - they are very cool people, help in carrying out tasks and encourage further work and not give up.
I hope that I will be able to win because it would open new development opportunities and future work for me - it would be easier to get the necessary qualifications.

Thanks to taking part in GCI, I was able to polish my English and work level based on how I work in large companies.I gained a lot of knowledge and knowledge in technology. I got used to using git, writing in English and searching for information, i.e. working with documentation (sometimes language, sometimes a website such as IBM Bluemix) and Stack Overflow. This contest was also an opportunity to leave the cave and test your skill
Thanks to JBoss Comunity, I can exchange my experiences and skills with other GCI participants.

I think that Google has done a great job creating this competition, thanks to which many young people have a chance for a better future.
If you're under 18, try Google Code-In, because it's really worth it. My advice: choose organizations with which fewer people work
Finally, I hope to improve my skills, and now I'm getting to work.